reorganize event list
